% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_beta}
\alias{gghm_beta}
\title{Plot beta posterior estimates as colored boxes}
\usage{
gghm_beta(
  Hm,
  order_x = NA,
  grouping_var_y = NA,
  grouping_var_y_2 = NA,
  n_gvy = 1,
  spp_exclude = NA,
  support_level = 0.89,
  lut_varnames = NULL,
  lut_sppnames = NULL,
  no_intercept = TRUE,
  title = NA
)
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{order_x}{order the variables on the x axis}

\item{grouping_var_y}{group the species on the Y axis by traits}

\item{grouping_var_y_2}{second grouping variable}

\item{n_gvy}{number of y axis grouping variables}

\item{spp_exclude}{a vector of species names to exclude from the figure}

\item{support_level}{the support level at which to make parameters visible (default = 0.89)}

\item{lut_varnames}{a named vector (look up table) for changing the names of environmental covariates}

\item{lut_sppnames}{a named vector (look up table) for changing the names of species}

\item{no_intercept}{exclude the intercept parameters from the plot?}

\item{title}{the plot title}
}
\description{
Plot beta posterior estimates as colored boxes
}
