% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_beta2}
\alias{gghm_beta2}
\title{Plot beta estimates using PDFs}
\usage{
gghm_beta2(
  Hm,
  order_var = "prevalence",
  grouping_var_y = NA,
  groupiin_var_y_2 = NA,
  lut_gensp = NA,
  excluded_spp = NA,
  included_variables = NA,
  lut_ivars = NA
)
}
\description{
Plot beta estimates using PDFs
}
\examples{
data("Hm")
gghm_beta2(Hm)

}
