% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_beta3}
\alias{gghm_beta3}
\title{Plot beta posterior estimates using point bars}
\usage{
gghm_beta3(
  Hm,
  group_by_trait = NA,
  included_variables = NA,
  lut_ivars = NA,
  top_x_species = NA,
  lut_gensp = NA
)
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{group_by_trait}{a column name from the trait data (Hm$TrData) with which to sort the y axis}

\item{included_variables}{a vector of environmental variable names to indicate which variables to include}

\item{lut_ivars}{a named vector of environmental variable names to change them.}

\item{top_x_species}{a numeric, indicating to only plot the top X most prevalent species}

\item{lut_gensp}{a named vector of species names to change them}
}
\description{
Plot beta posterior estimates using point bars
}
\examples{
data("Hm")
gghm_beta3(Hm)

}
