% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_beta_tests}
\alias{gghm_beta_tests}
\title{Do tests on parameters}
\usage{
gghm_beta_tests(Hm, ci = 0.95)
}
\arguments{
\item{Hm}{Hmsc object}

\item{ci}{Credible interval}
}
\description{
Do tests on parameters
}
