% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_convergence}
\alias{gghm_convergence}
\title{Plot model convergence indexes}
\usage{
gghm_convergence(
  Hm,
  beta = TRUE,
  V = FALSE,
  gamma = FALSE,
  omega = FALSE,
  title = "Model Convergence"
)
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{beta}{if TRUE, plots the beta parameters (environmental covariates).}

\item{V}{if TRUE, plots the V parameters.}

\item{gamma}{if TRUE, plots the gamma parameters (traits).}

\item{omega}{if TRUE, plots the omega parameters (spp associations).}

\item{title}{character string.}
}
\description{
Plot model convergence indexes
}
