% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_fit}
\alias{gghm_fit}
\title{get Tjur R2 for each species}
\usage{
gghm_fit(Hm, which = "r2", sp_names = "none", title = "Variance Explained")
}
\arguments{
\item{which}{either all, r2 or named. all is a histogram with R2 and RSME. r2 is just R2. named is bar plot each species named on the  y axis.}
}
\description{
get Tjur R2 for each species
}
