% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_gamma}
\alias{gghm_gamma}
\title{plot trait covariate relationships}
\usage{
gghm_gamma(
  Hm,
  support_level = 0.89,
  no_intercept = TRUE,
  title = "Effects on Traits"
)
}
\arguments{
\item{Hm}{an Hmsc object}

\item{support_level}{make the boxes dissappear below this number}

\item{no_intercept}{exclude the intercept?}

\item{title}{plot title}
}
\description{
plot trait covariate relationships
}
