% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_gamma2}
\alias{gghm_gamma2}
\title{plot trait covariate associations}
\usage{
gghm_gamma2(Hm, lut_varnames = NA)
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{lut_varnames}{a named vector used to change the variable names}
}
\description{
plot trait covariate associations
}
