% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_gradient_c}
\alias{gghm_gradient_c}
\title{gradient plots for continuous variables}
\usage{
gghm_gradient_c(Hm, focal_variable, fn = NA, title = NA, chains)
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{focal_variable}{the variable against which to plot model predictions}

\item{fn}{filename if you wish to save (optional)}

\item{title}{character. the plot title (optional)}

\item{chains}{number of MCMC chains used}
}
\description{
gradient plots for continuous variables
}
