% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_r2_tjur}
\alias{gghm_r2_tjur}
\title{get Tjur R2 for each species}
\usage{
gghm_r2_tjur(Hm, title = "Variance Explained", sp_names = "none")
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{title}{the plot title}

\item{sp_names}{look up table to change species names to be more presentable}
}
\description{
get Tjur R2 for each species
}
