% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_traceplot}
\alias{gghm_traceplot}
\title{Trace plots}
\usage{
gghm_traceplot(Hm, which = "beta")
}
\arguments{
\item{Hm}{Hmsc object}

\item{which}{Can be "beta", "gamma", or "v"}
}
\description{
Trace plots
}
