% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_vp}
\alias{gghm_vp}
\title{Plot variance partitioning}
\usage{
gghm_vp(
  Hm,
  title = "Variance Explained",
  cols = NULL,
  lut_varnames = NULL,
  lut_sppnames = NULL
)
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{title}{the title of the plot}

\item{cols}{a vector of colors}

\item{lut_varnames}{a named vector (look up table) for changing the names of variables. Useful for making things more presentable for publication.}

\item{lut_sppnames}{a named vector (look up table) for changing the names of species}
}
\description{
Plot variance partitioning
}
\examples{
data('Hm')
gghm_vp(Hm)

}
