% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{gghmsc_beta2}
\alias{gghmsc_beta2}
\title{Plot beta estimates using PDFs}
\usage{
gghmsc_beta2(Hm, lut_gensp = NA, included_variables = NA, lut_ivars = NA)
}
\description{
Plot beta estimates using PDFs
}
