% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{gghmsc_gamma}
\alias{gghmsc_gamma}
\title{plot trait covariate relationships}
\usage{
gghmsc_gamma(
  Hm,
  support_level = 0.89,
  no_intercept = TRUE,
  title = "Effects on Traits"
)
}
\description{
plot trait covariate relationships
}
