% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{gghmsc_omega}
\alias{gghmsc_omega}
\title{create a correlation plot for species associations}
\usage{
gghmsc_omega(
  Hm,
  support_level = 0.89,
  hc.method = "single",
  hc.order = TRUE,
  lut_gensp,
  axis_text_colors_x = "black",
  axis_text_colors_y = "black",
  title = "Residual Species Associations"
)
}
\description{
create a correlation plot for species associations
}
