% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{gghmsc_plot}
\alias{gghmsc_plot}
\title{Trace plots}
\usage{
gghmsc_plot(Hm, which = "beta")
}
\arguments{
\item{Hm}{Hmsc object}

\item{which}{Can be "beta", "gamma", or "v"}
}
\description{
Trace plots
}
