% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{gghmsc_vp}
\alias{gghmsc_vp}
\title{Plot variance partitioning}
\usage{
gghmsc_vp(
  Hm,
  title = "Variance Explained",
  cols = NULL,
  lut_varnames = NULL,
  lut_sppnames = NULL
)
}
\description{
Plot variance partitioning
}
