% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{ggplot_beta}
\alias{ggplot_beta}
\title{Plot beta posterior estimates as colored boxes}
\usage{
ggplot_beta(
  Hm,
  grouping_var = NA,
  support_level = 0.89,
  lut_varnames = NULL,
  lut_sppnames = NULL,
  no_intercept = TRUE,
  title = NA
)
}
\description{
Plot beta posterior estimates as colored boxes
}
