% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{ggplot_beta2_drake}
\alias{ggplot_beta2_drake}
\title{Plot beta estimates using PDFs}
\usage{
ggplot_beta2_drake(Hm, lut_gensp = NA, included_variables = NA, lut_ivars = NA)
}
\description{
Plot beta estimates using PDFs
}
