% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsc_ggplotting.R
\name{ggplot_ess}
\alias{ggplot_ess}
\title{Plot effective sample size}
\usage{
ggplot_ess(
  Hm,
  beta = TRUE,
  V = FALSE,
  gamma = FALSE,
  omega = FALSE,
  title = "Model Convergence"
)
}
\description{
Plot effective sample size
}
