% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hm}
\alias{Hm}
\title{An Hmsc object}
\format{
\subsection{'Hm'}{

An Hmsc object. Basically a list.
\describe{
\item{XData}{data frame of environmental covariates}
\item{Y}{data frame of species abundances or occurrences}
\item{TrData}{data frame of trait data}
...
}
}
}
\source{
\url{https://www.github.com/admahood/time-since-fire}
}
\usage{
Hm
}
\description{
An Hmsc object
}
\keyword{datasets}
