% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_beta}
\alias{gghm_beta}
\title{Plot beta posterior estimates as colored boxes}
\usage{
gghm_beta(
  Hm,
  order_x = NA,
  grouping_var_y = NA,
  grouping_var_y_2 = NA,
  n_gvy = 1,
  spp_exclude = NA,
  support_level = 0.89,
  lut_varnames = NULL,
  lut_sppnames = NULL,
  no_intercept = TRUE,
  title = NA
)
}
\arguments{
\item{order_x}{order the variables on the x axis}

\item{grouping_var_y}{group the species on the Y axis by traits}
}
\description{
Plot beta posterior estimates as colored boxes
}
\examples{
data("Hm")
gghm_beta(Hm)


}
