% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_beta2}
\alias{gghm_beta2}
\title{Plot beta estimates using PDFs}
\usage{
gghm_beta2(
  Hm,
  order_var = "prevalence",
  grouping_var_y = NA,
  grouping_var_y_2 = NA,
  lut_gensp = NA,
  excluded_spp = NA,
  included_variables = NA,
  top_x_species = NA,
  lut_ivars = NA,
  group_by_trait = NA
)
}
\arguments{
\item{Hm}{an Hmsc model object}

\item{order_var}{order the variables on the y axis with this trait variable}

\item{grouping_var_y}{group the species on the Y axis by traits}

\item{grouping_var_y_2}{second grouping variable}

\item{lut_gensp}{a named vector of species names to change them}

\item{excluded_spp}{a vector of species names to exclude from the figure}

\item{included_variables}{a vector of environmental variable names to indicate which variables to include}

\item{top_x_species}{a numeric, indicating to only plot the top X most prevalent species}

\item{lut_ivars}{a named vector of environmental variable names to change them.}

\item{group_by_trait}{UNDER CONSTRUCTION: a column name from the trait data (Hm$TrData) with which to sort the y axis}
}
\description{
Plot beta estimates using PDFs
}
\examples{
data("Hm")
gghm_beta2(Hm)

}
