% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_ess}
\alias{gghm_ess}
\title{Plot effective sample size}
\usage{
gghm_ess(
  Hm,
  beta = TRUE,
  V = FALSE,
  gamma = FALSE,
  omega = FALSE,
  title = "Model Convergence"
)
}
\description{
Plot effective sample size
}
