% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_omega}
\alias{gghm_omega}
\title{create a correlation plot for species associations}
\usage{
gghm_omega(
  Hm,
  support_level = 0.89,
  hc.method = "single",
  hc.order = TRUE,
  lut_gensp = NA,
  axis_text_colors_x = "black",
  axis_text_colors_y = "black",
  title = "Residual Species Associations"
)
}
\description{
create a correlation plot for species associations
}
