% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_r2_tjur}
\alias{gghm_r2_tjur}
\title{get Tjur R2 for each species}
\usage{
gghm_r2_tjur(Hm, title = "Variance Explained", sp_names = "none")
}
\arguments{
\item{which}{either all, r2 or named. all is a histogram with R2 and RSME. r2 is just R2. named is bar plot each species named on the  y axis.}
}
\description{
get Tjur R2 for each species
}
