% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gghm_vp}
\alias{gghm_vp}
\title{Plot variance partitioning}
\usage{
gghm_vp(
  Hm,
  title = "Variance Explained",
  cols = NULL,
  lut_varnames = NULL,
  lut_sppnames = NULL
)
}
\description{
Plot variance partitioning
}
