% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hm}
\alias{Hm}
\title{An Hmsc object}
\format{
\subsection{'Hmsc object'}{

An Hmsc object
}
}
\source{
\url{https://www.github.com/admahood/time-since-fire}
}
\usage{
Hm
}
\description{
Note: Some sites have more than one polygon. There are 59 polygons and 47 total sites.
}
\keyword{datasets}
